Share resources initialization
